create shared folder, move /functions /hooks /img there
