+++ src/components/Data/UserPercentiles/PercentileDisplay.scss
@@ -17,7 +17,7 @@
   &__graph {
     position: relative;
     padding-bottom: 100%;
-    background-image: url('../../../img/grid-only.svg');
+    background-image: url('../../../shared/img/grid-only.svg');
     background-position: center;
     background-repeat: no-repeat;
     background-size: auto;
