+++ src/components/Data/d3/drawTastyBox.js
@@ -101,7 +101,9 @@ const drawTastyBox = () => {
       .data(data)
       .enter()
       .append('svg:image')
-      .attr('xlink:href', (d) => require(`../../../img/${d.name}.min.svg`))
+      .attr('xlink:href', (d) =>
+        require(`../../../shared/img/${d.name}.min.svg`)
+      )
       .attr('x', (d) => x(d.name) - imgSize / 2)
       .attr('y', (d) => y(d.avg) - imgSize / 2)
       .attr('width', imgSize)
