+++ src/components/Footer.js
@@ -2,10 +2,10 @@ import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
 import toaster from 'toasted-notes';
 import ReactGA from 'react-ga';
-import Button from './Button';
+import Button from '../shared/components/Button';
 import SubmitModal from './SubmitModal';
 import XKCDModal from './XKCDModal';
-import roundRatings from '../functions/roundRatings';
+import roundRatings from '../shared/functions/roundRatings';
 import 'react-responsive-modal/styles.css';
 import 'toasted-notes/src/styles.css';
 import '../components/Toasts.scss';
