+++ src/components/Fruit.js
@@ -17,7 +17,7 @@ const Fruit = ({
   const nodeRef = useRef(null);
   const [isDraggingOverGraph, setIsDraggingOverGraph] = useState(null);
   const marginForOffGraphFruits = scale.imgSize * 0.25; // scales responsively
-  const src = require(`../img/${name}.min.svg`);
+  const src = require(`../shared/img/${name}.min.svg`);
 
   // Use rating to determine if the fruit is currently placed on the graph
   let isOnGraph = !ratings[name] ? false : true;
