+++ src/components/Graph.scss
@@ -5,7 +5,7 @@
     width: 100%;
     max-width: 65vh;
     margin: 0 auto;
-    background-image: url('../img/axis.svg');
+    background-image: url('../shared/img/axis.svg');
     background-position: center;
     background-repeat: no-repeat;
     background-size: auto;
