+++ src/components/Header.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link, useHistory } from 'react-router-dom';
-import Button from './Button';
+import Button from '../shared/components/Button';
 import './Header.scss';
 
 const Header = () => {
