+++ src/shared/functions/cleanFruitName.js
